// src/app/models/opening.model.ts
export interface Opening {
    id: string;
    posId: string;  
    counter: number;
    profit: number;
    initialBase: number;
    total: number;    
    posCounter: number;
    salesCounter: number;    
    cashierId: string;
    openingDate: string;
    openingHour: string;
    closingDate: string;
    closingHour: string;    
}